+++ src/source/canvas.cpp
@@ -1,6 +1,6 @@
 #include "../headers/canvas.h"
 
-Vector3 canvas_to_viewport(const Canvas& c, const Viewport& v, int x, int y, int d) {
+Vector3 canvas_to_viewport(const Canvas& c, const Viewport& v, float x, float y, float d) {
 	float vx = x * (float(v.width) / float(c.width));
 	float vy = y * (float(v.height) / float(c.height));
 	return Vector3(vx, vy, d);
